refactor: six files crossed the 500-LoC cap in this stack — split them
Four crossed it in the stacked merges and two in this branch's own work. Each
split is an exact piecewise copy at a seam the file already had; no behaviour
and no test changed in any move.

* concurrency_tests.rs 589 -> 317 + concurrency_verdict_tests.rs. The
  run-verdict half was already a separate concern; the ladder widening to
  C=128 roughly doubled the rung tables both halves assert on.
* bench_tests.rs 526 -> 306 + bench_serve_pin_tests.rs. The serve-pin tests
  are the half that says what a gate LAUNCHES, which grew a second subject.
* cuda_backend.rs 501 -> 388 + cuda_backend/alloc_ledger.rs. The allocation
  ledger from #650 is a coherent unit: what is live, how big, which call site.
  One correction carried in the move — the doc comment above `record_alloc`
  described the `registry` field, not the method, on main.
* dflash_loader.rs 534 -> 422 + dflash_loader/loader_tests.rs.
* lora/types.rs 506 -> 373 + types_slots.rs. The pool LAYOUT stays in types.rs;
  the half that mutates which adapter is resident moves out.
* qwen3_ssm/init.rs 501 -> 488 + init_kernels.rs. This one crossed by a single
  line (`lora_out_proj: None`), so the conditional kernel-lookup helper moves
  rather than the constructor being carved up for one line.

No .rs file outside the workflow's allow-list now exceeds 500. Workspace check,
clippy and fmt clean; 876 atlas-plugin tests pass with the same three
unmeasured-entry bootstrap failures as before.
