+++ crates/spark-model/src/layers/qwen3_ssm/init_kernels.rs
@@ -1,0 +1,25 @@
+// SPDX-License-Identifier: AGPL-3.0-only
+
+//! Kernel lookups the SSM layer's constructor makes conditionally.
+//!
+//! Split from `init.rs` for the 500-LoC cap, which the file crossed by one
+//! line when the LoRA `out_proj` slot joined the layer. Exact piecewise copy.
+
+use super::*;
+
+/// Resolve one `hyper_connection` entry point, but ONLY for a model that
+/// carries the highway. Skipping the lookup rather than discarding its result
+/// is the point: an un-issued lookup leaves no failed row in the fail-closed
+/// startup audit, so what remains there is what someone has to act on.
+#[track_caller]
+pub(super) fn hc_kernel(
+    config: &atlas_core::config::ModelConfig,
+    gpu: &dyn GpuBackend,
+    func: &str,
+) -> KernelHandle {
+    if config.hc_mult > 0 {
+        crate::layers::try_kernel(gpu, "hyper_connection", func)
+    } else {
+        KernelHandle(0)
+    }
+}
